+++ src/Utils/Auth.js
@@ -54,14 +54,26 @@ class Auth {
   /**
      * @description Resolves a promisse if the token is valid following LTI 1.3 standards.
      * @param {String} token - JWT token to be verified.
+     * @param {String} state - State validation parameter.
+     * @param {Object} validationParameters - Stored validation parameters retrieved from cookies.
      * @param {Function} getPlatform - getPlatform function to get the platform that originated the token.
      * @param {String} ENCRYPTIONKEY - Encription key.
      * @returns {Promise}
      */
-  static async validateToken (token, getPlatform, ENCRYPTIONKEY, logger, Database) {
+  static async validateToken (token, state, validationParameters, getPlatform, ENCRYPTIONKEY, logger, Database) {
     const decodedToken = jwt.decode(token, { complete: true })
     const kid = decodedToken.header.kid
     const alg = decodedToken.header.alg
+
+    provAuthDebug('Attempting to validate iss claim')
+    provAuthDebug('Request Iss claim: ' + validationParameters.iss)
+    provAuthDebug('Response Iss claim: ' + decodedToken.payload.iss)
+    if (!validationParameters.iss || (validationParameters.iss !== decodedToken.payload.iss)) throw new Error('IssClaimDoesNotMatch')
+
+    provAuthDebug('Attempting to validate state')
+    provAuthDebug('Request state: ' + validationParameters.state)
+    provAuthDebug('Response state: ' + state)
+    if (!validationParameters.state || (validationParameters.state !== state)) throw new Error('StateClaimDoesNotMatch')
 
     provAuthDebug('Attempting to retrieve registered platform')
     const platform = await getPlatform(decodedToken.payload.iss, ENCRYPTIONKEY, logger, Database)